requirements.txt
absl-py
alphashape
ezdxf
shapely
sexpdata
